this is confusing
https://gamebanana.com/threads/218927 section 3.7

I don't understand fully, which one is that one that? there's alot of the mods including other mods without permission (presumably) like https://gamebanana.com/wips/55046 and is okay. btw that's wip page, not mod page. but it has the real mod page to Gamejolt due to compression failure. anyway

c'mon, we also inserted these mods and will go further while making sure everyone yoinked got credited as well as their name. are we absolutely allowed to yoink, or absolutely not? Please, you must be specific on what offender you target.

continued the GF Home theather. hoof! this is whole of a chore! can't there a technology to ease these?! right now, the existing technology https://gamebanana.com/tools/7032 only covers BF at the moment. We have a bit of knowledge of Texture Atlas XML now, so we tried to go as far as we can.

whoops! for these some things!
- forgot to check cpp on one part of Discord RPC
- curstage!!! red!!! yess.

we've build README on top of the upstream one. all in one page. you can now read glympse of information what is this mod about.

1 JUNI GOOGLE WILL KILL UNLIMITED PHOTO STORAGE!!! After that that, photo uploaded eats your Google Drive quota. that's now all useless. no matter what compression type High or Low.