[skip ci] polgromatd
whups misfire. put the sm files back to source assets pls!

fortritri because I thought the error was because number is not allowed, turns out we had null in playable note parts. this crashes when that number of playable notes being queried in FreePlayState. Because, the array is empty but they attempted to change that first note parameter that doesn't exist in!!! yep, I fixed it by forcing it says 0.0 float.

pls finish gf-covid, we're close to next month already!