[skip ci] oh no error
don't add || at the end of the show in the if. Blammed mod hscript guys

forgot to force Hscript

adfasdfasdfasdfasdfasdfasdfasdfasd

skriping the mobile. Remember, the Android loading is different, so in Freeplay we also skip that one up here.

attempt to path Asset2File only right when bitmap data in Mobile.