[skip ci] Sapi Polen 17
Finally rember to make separated mods folder that are not going to be included in the main LFM unless compiled debug or defined `FORCE_INCLUDE_DLC`.

Whoah! OMG, there's a lot of formats I got to choose. Well, if nobody reconciliate, it will be bewildering which framework forks to choose then. This is scary. NO, not meaning competition is bad, I just... fear, alot of different formats just to achieve 1 thing, That's brain overload!!!

Engima Engine https://github.com/EnigmaEngine/EnigmaEngine

You also would like to not support Texted Week data because sometimes the order no matter what could be messed up, and it will be difficult to track which sub mod is the culprit. JSON granular week again, turns out to be the best. Stable, and definite what's this week is about. unlike that Texted which had separate and unstable which defines which.