[skip ci] dialog boksu
Improve pie dial for auto click timer status. you see that if the timer isn't started, there are residue of reduction, and yeah, comparing time set & left of the FlxTimer does not completely complies 1.0. seriously, what bug is this? so peck it. let me full it myself instead. if the Autoclick is ON but timer not start, it shall be 1.0, if started it shall be whatever time left per time set. if Autoclick is OFF, it shall be 0.0

NEW Modchart function! do function when Dialog opens, ended, and skipped.

pls don't forget templateoid! Modchart templates