[skip ci] oh man cartoon corner dot UNTESTED
here cartoon corner dot that appears at the end of the episode to mark the end, then transisition to the end card right at the edge end of the film strip yeah!

Now here, it appears when you finished all notes, then disappear after a second or few. after that, it is up to whether the music file finished, OR if configured, end earlier. like usual.