[skip ci] ontrep
we got to fix alot of things we forgot to transmigrate. these character I had, I forgot to transfer to a new format.

also woops, forgor Botplay text psychfication! what else we forgot?

FINALLY THE HEARTBEAT FETISH IS PECKING WORKING!!!
now need the way to steth, and the sound yess.