[skip ci] trabel
the PC build budget exploded around IDR 40 Million. And our financial situation makes it could not reach this for even. Well this is for the old version right before AM5 AMD. Not to mention, it gonna be even higher. I plan to start from here and then throw away replace better later, idk.